Fix things so that libmach can build.

+++ sys/include/libc.h
@@ -11,7 +11,7 @@
 #pragma	src	"/sys/src/libc"
 
 #define	nelem(x)	(sizeof(x)/sizeof((x)[0]))
-#define	offsetof(s, m)	(uint32_t)(&(((s*)0)->m))
+#define	offsetof(s, m)	(uintptr_t)(&(((s*)0)->m))
 #define	assert(x)	if(x){}else _assert("x")
 
 /*
